Makes test.py Python2.7 compatible.
Importing print_function and unicode_literals features from
__future__.

Adding character encoding declaration to top of file.

Check version in mkdir_p_and_open to handle changes in the os module
from version 2.7 to 3+.

+++ test.py
@@ -1,9 +1,14 @@
+# This Python file uses the following encoding: utf-8
+
+from __future__ import print_function
+from __future__ import unicode_literals
 
 from pysbolgraph.SBOL2Graph import SBOL2Graph
 from pysbolgraph.terms import Biopax, SBOL2
 from glob import glob
 import json
 import os
+from sys import version_info
 import requests
 
 
@@ -15,8 +20,18 @@ def load(f):
 
 
 def mkdir_p_and_open(filename):
-    os.makedirs(filename.rpartition('/')[0], exist_ok=True)
-    return open(filename, "wb+")
+    if version_info > (3, 0):
+        os.makedirs(filename.rpartition('/')[0], exist_ok=True)
+        return open(filename, "wb+")
+    else:
+        try:
+            os.makedirs(filename.rpartition('/')[0])
+            return open(filename, "wb+")
+        except OSError as e:
+            if e.errno == os.errno.EEXIST:
+                return open(filename, "wb+")
+            else:
+                raise e
 
 
 files = glob('SBOLTestSuite/SBOL2/*.xml')
